+++ src/migrations/M240819124324Tenant.php
@@ -37,14 +37,6 @@ class M240819124324Tenant extends Migration
             'id',
             'SET NULL'
         );
-
-        $tenant = Tenant::create();
-        $tenant->loadDefaultValues();
-        $tenant->name = 'Default';
-        $tenant->url = 'https://www.example.com/';
-        $tenant->language = Yii::$app->language;
-
-        $tenant->insert(false);
     }
 
     public function safeDown(): void